Delete BuildingView, rename views and add dummy view content.
