+++ src/views/BookingsView.vue
