--- src/views/RoomView.vue
@@ -1,15 +1,0 @@
-<template>
-  <BaseCard title="RoomView"/>
-</template>
-
-<script>
-import BaseCard from "@/components/card/BaseCard";
-export default {
-  name: "RoomView",
-  components: {BaseCard}
-}
-</script>
-
-<style scoped>
-
-</style>
